requirements.txt
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
